% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Compute_Cn1n2}
\alias{Compute_Cn1n2}
\title{Compute the Delta of the dynamic programming}
\usage{
Compute_Cn1n2(x)
}
\arguments{
\item{x}{the matrix of rank}
}
\description{
Compute the Delta of the dynamic programming in \code{\link{Rcpp}}
}
