% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuChPoint_Class.R
\docType{methods}
\name{print,MuChPoint-method}
\alias{print,MuChPoint-method}
\title{Print for the class of object returned by the \code{MuChPoint} function.}
\usage{
\S4method{print}{MuChPoint}(x, N = NULL)
}
\arguments{
\item{x}{an object with class \code{MuChPoint}}

\item{N}{a numeric between 1 and length(x@N) for the number of break-points desired.}
}
\description{
Print for the class of object returned by the \code{MuChPoint} function.
}
