% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuChPoint_Class.R
\docType{methods}
\name{summary,MuChPoint-method}
\alias{summary,MuChPoint-method}
\title{Summary of a \code{MuChPoint} object.}
\usage{
\S4method{summary}{MuChPoint}(object)
}
\arguments{
\item{object}{an object of class \code{MuChPoint}.}
}
\description{
Summary of a \code{MuChPoint} object.
}
\examples{
require(MuChPoint)
mu=c(rep(c(rep(1,25),rep(0,25)),3))\%*\%t(rep(c(rep(0,25),rep(1,25)),3))
Y=matrix(rnorm(150^2,0,2),150)+mu+t(mu)
Y=as.matrix(Matrix::forceSymmetric(Y))
res=MuChPoint(Y)
summary(res)

}
\seealso{
\code{\linkS4class{MuChPoint}}.
}
